+++ spec/kitchen_spec/directions/bake_further_research_spec.rb
@@ -40,6 +40,27 @@ RSpec.describe Kitchen::Directions::BakeFurtherResearch do
     )
   end
 
+  let(:chapter_with_module_with_titles_children) do
+    chapter_element(
+      <<~HTML
+        <div data-type='page' id="00" class="introduction">
+          <h1 data-type="document-title" itemprop="name" id="intro">Introduction page!</h1>
+          <section class="further-research" data-element-type="further-research">
+              <h3 data-type='title'>Kitchen Prep</h3>
+              <p>This moves too!</p>
+          </section>
+        </div>
+        <div data-type='page' id ="01">
+          <h1 data-type="document-title" itemprop="name" id="first"><em data-effect="italics">Italics</em> & not italics with <sub>subscript</sub> and <sup>superscript</sup> text First Title</h1>
+          <section class="further-research" data-element-type="further-research">
+              <h3 data-type='title'>Roasting</h3>
+              <p>Many paragraphs provide a good summary.</p>
+          </section>
+        </div>
+      HTML
+    )
+  end
+
   context 'when v1 is called on a chapter' do
     it 'works' do
       metadata = metadata_element.append(child:
@@ -50,6 +71,20 @@ RSpec.describe Kitchen::Directions::BakeFurtherResearch do
       described_class.v1(chapter: chapter, metadata_source: metadata)
       expect(
         chapter
+      ).to match_snapshot_auto
+    end
+  end
+
+  context 'when v1 is called on a chapter with module with titles children' do
+    it 'keeps the module titles children while creating section title' do
+      metadata = metadata_element.append(child:
+        <<~HTML
+          <div data-type="random" id="subject">Random - should not be included</div>
+        HTML
+      )
+      described_class.v1(chapter: chapter_with_module_with_titles_children, metadata_source: metadata)
+      expect(
+        chapter_with_module_with_titles_children
       ).to match_snapshot_auto
     end
   end